recon_tools/requirements.txt
dnspython>=2.3.0
python-whois>=0.8.0
requests>=2.28.0
colorama>=0.4.6
urllib3>=1.26.0